+++ Frameworks/OmniAppKit/OpenStepExtensions.subproj/NSTableView-OAExtensions.h
@@ -40,7 +40,7 @@ typedef enum _OATableViewRowVisibility {
 @interface NSObject (NSTableViewOAExtendedDataSource)
 
 // Searching
-- (BOOL)tableView:(NSTableView *)tableView itemAtRow:(int)row matchesPattern:(id <OAFindPattern>)pattern;
+- (BOOL)tableView:(NSTableView *)tableView itemAtRow:(NSInteger)row matchesPattern:(id <OAFindPattern>)pattern;
     // Implement this if you want find support.
 - (NSTableColumn *)tableViewTypeAheadSelectionColumn:(NSTableView *)tableView;
     // Return non-nil to enable type-ahead selection. Needs a column whose values are strings (or respond to -stringValue)... presumably the names of your rows' represented objects. If your table has only one column, we'll choose it by default unless you implement this method to return nil.
@@ -54,7 +54,7 @@ typedef enum _OATableViewRowVisibility {
 // Drag image control
 - (NSArray *)tableViewColumnIdentifiersForDragImage:(NSTableView *)tableView;
     // If you have a table similar to a Finder list view, where one or more columns contain a representation of the object associated with each row, and additional columns contain supplemental information (like sizes and mod dates), implement this method to specify which column(s) should be part of the dragged image. (Because you want to show the user that you're dragging a file, not a file and a date and a byte count.)
-- (BOOL)tableView:(NSTableView *)tableView shouldShowDragImageForRow:(int)row;
+- (BOOL)tableView:(NSTableView *)tableView shouldShowDragImageForRow:(NSInteger)row;
     // If you'd like to support dragging of multiple-row selections, but want to control which of the selected rows is valid for dragging, implement this method in addition to -tableView:writeRows:toPasteboard:. If none of the selected rows are valid, return NO in -tableView:writeRows:toPasteboard:. If some of them are, write the valid ones to the pasteboard and return YES in -tableView:writeRows:toPasteboard:, and implement this method to return NO for the invalid ones. This prevents them from being drawn as part of the drag image, so that the items the user appears to be dragging are in sync with the items she's actually dragging.
 
 - (NSDragOperation)tableView:(NSTableView *)tableView draggingSourceOperationMaskForLocal:(BOOL)isLocal;
@@ -66,7 +66,7 @@ typedef enum _OATableViewRowVisibility {
     // Normally tables like to move you to the next row when you hit return after editing a cell, but that's not always desirable.
 
 // Context menus
-- (NSMenu *)tableView:(NSTableView *)tableView contextMenuForRow:(int)row column:(int)column;
+- (NSMenu *)tableView:(NSTableView *)tableView contextMenuForRow:(NSInteger)row column:(NSInteger)column;
 
 @end
