+++ Frameworks/OmniAppKit/Widgets.subproj/OATextWithIconCell.m
@@ -179,7 +179,7 @@ NSString *OATextWithIconCellImageKey = @"image";
     [super editWithFrame:aRect inView:controlView editor:textObj delegate:anObject event:theEvent];
     _oaFlags.settingUpFieldEditor = NO;
 }
-- (void)selectWithFrame:(NSRect)aRect inView:(NSView *)controlView editor:(NSText *)textObj delegate:(id)anObject start:(int)selStart length:(int)selLength;
+- (void)selectWithFrame:(NSRect)aRect inView:(NSView *)controlView editor:(NSText *)textObj delegate:(id)anObject start:(NSInteger)selStart length:(NSInteger)selLength;
 {
     _calculateDrawingRectsAndSizes;
