+++ Frameworks/OmniInspector/OIInspectorWindow.m
@@ -20,7 +20,7 @@ RCS_ID("$Id$");
 
 @implementation OIInspectorWindow
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag;
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag;
 {
     if (![super initWithContentRect:contentRect styleMask:aStyle backing:bufferingType defer:flag])
         return nil;
